More hompage dropdowns / customer requests
Final styling for dropdowns possible to this point with the data I have
so far.
Added gradient to background, but may need to change it for app preview
because it is repeated improperly.
MP Controls pagenation dropdown adjusted.
Search and phone number/social icons in header set to float so they are
still proportionate when zooming the view in a browser.