clickable options for size
Set this up so that if a single option group is present, those options
will appear in the product list as clickable (not working in results
yet). If we KNOW inventory is available, they'll add to the cart. if we
know inventory is NOT available, they'll fade out some.  If inventory is
unknown, they'll click thru to the product detail page.