Recently viewed onDeparts
Addition of recently viewed item was moved from quickstart to custom
extension to allow it to happen when the prod page was left, rather than
when it was loaded. This allows the product list to be displayed on
multiple pages, and prevents prods from being loaded on the first visit
to their own page.
Still needs empty list provision, and possibly integration w/ search or
quickview.